project/requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
appdirs==1.4.4
async-timeout==5.0.1
asyncssh==2.22.0
atpublic==7.0.0
attrs==26.1.0
av==17.0.0
billiard==4.2.4
bitsandbytes==0.49.2
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
cryptography==43.0.3
cuda-bindings==13.2.0
cuda-pathfinder==1.5.2
cuda-toolkit==13.0.2
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==1.1.0
dvc==3.67.1
dvc-data==3.18.3
dvc-gdrive==3.0.1
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
einops==0.8.2
entrypoints==0.4
exceptiongroup==1.3.1
filelock==3.25.2
flatten-dict==0.4.2
flufl-lock==9.0.0
frozenlist==1.8.0
fsspec==2026.3.0
funcy==2.0
gitdb==4.0.12
gitpython==3.1.46
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
googleapis-common-protos==1.74.0
grandalf==0.8
gto==1.9.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface-hub==0.36.2
hydra-core==1.3.2
idna==3.11
iterative-telemetry==0.0.10
jinja2==3.1.6
kombu==5.6.2
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.4.2
numpy==2.2.6
nvidia-cublas==13.1.0.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti==13.0.85
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime==13.0.96
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver==12.0.4.66
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse==12.6.3.3
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.21.5
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
nvidia-nvtx-cu12==12.4.127
oauth2client==4.1.3
omegaconf==2.3.0
orjson==3.11.8
packaging==26.0
pathspec==1.0.4
pillow==12.2.0
platformdirs==4.9.4
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.2.2
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycocotools==2.0.11
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pydot==4.0.1
pydrive2==1.21.3
pygit2==1.18.2
pygments==2.20.0
pygtrie==2.5.0
pyopenssl==24.2.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyyaml==6.0.3
qwen-vl-utils==0.0.14
regex==2026.4.4
requests==2.33.1
rich==14.3.3
rsa==4.9.1
ruamel-yaml==0.19.1
safetensors==0.7.0
scmrepo==3.6.2
semver==3.0.4
sentencepiece==0.2.1
sentry-sdk==2.57.0
setuptools==81.0.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
smmap==5.0.3
sqltrie==0.11.2
sympy==1.13.1
tabulate==0.10.0
timm==1.0.26
tokenizers==0.21.4
tomlkit==0.14.0
torch==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.3
transformers==4.49.0
triton==3.2.0
typer==0.24.1
typing==3.10.0.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
voluptuous==0.16.0
wandb==0.25.1
wcwidth==0.6.0
yarl==1.23.0
zc-lockfile==4.0
